+++ app/controllers/twinkle/appcast_controller.rb
@@ -1,7 +1,7 @@
 module Twinkle
   class AppcastController < ApplicationController
     def show
-      @app = App.with_versions.find_by!(slug: params[:slug])
+      @app = App.includes(:versions).find_by!(slug: params[:slug])
       Event.create(app: @app, **event_params)
       render layout: false, formats: :xml
     end