+++ test/models/version_test.rb
@@ -5,7 +5,7 @@ class VersionTest < ActiveSupport::TestCase
     version = Twinkle::Version.new(published: true)
     assert_not version.save
     assert_equal ["can't be blank"], version.errors[:number]
-    assert_equal ["can't be blank"], version.errors[:build]
+    assert_equal ["is not a number"], version.errors[:build]
     assert_equal ["can't be blank"], version.errors[:description]
     assert_equal ["can't be blank"], version.errors[:binary_url]
     assert_equal ["can't be blank"], version.errors[:length]
@@ -15,7 +15,7 @@ class VersionTest < ActiveSupport::TestCase
     version = Twinkle::Version.new(published: false)
     assert_not version.save
     assert_equal ["can't be blank"], version.errors[:number]
-    assert_equal ["can't be blank"], version.errors[:build]
+    assert_equal ["is not a number"], version.errors[:build]
     assert_equal [], version.errors[:description]
     assert_equal [], version.errors[:binary_url]
     assert_equal [], version.errors[:length]
@@ -25,6 +25,18 @@ class VersionTest < ActiveSupport::TestCase
     version = Twinkle::Version.new(binary_url: "not-a-url")
     assert_not version.save
     assert_equal ["is not a valid URL"], version.errors[:binary_url]
+  end
+
+  test "versions require a valid URL for the release_notes_link" do
+    version = Twinkle::Version.new(release_notes_link: "not-a-url")
+    assert_not version.save
+    assert_equal ["is not a valid URL"], version.errors[:release_notes_link]
+  end
+
+  test "versions require a valid URL for the full_release_notes_link" do
+    version = Twinkle::Version.new(full_release_notes_link: "not-a-url")
+    assert_not version.save
+    assert_equal ["is not a valid URL"], version.errors[:full_release_notes_link]
   end
 
   test "versions require a positive integer for the length" do
@@ -43,5 +55,17 @@ class VersionTest < ActiveSupport::TestCase
     version = Twinkle::Version.new(binary_url: "https://example.com")
     assert_not version.save
     assert_equal [], version.errors[:binary_url]
+  end
+
+  test "versions accept a valid URL for the release_notes_link" do
+    version = Twinkle::Version.new(release_notes_link: "https://example.com")
+    assert_not version.save
+    assert_equal [], version.errors[:release_notes_link]
+  end
+
+  test "versions accept a valid URL for the full_release_notes_link" do
+    version = Twinkle::Version.new(full_release_notes_link: "https://example.com")
+    assert_not version.save
+    assert_equal [], version.errors[:full_release_notes_link]
   end
 end