+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0
+
+- Changes build to an integer so ordering works properly on every database
+- Adds a lot of Sparkle 2.0 fields to the model
+
 ## 0.1.1
 
 - Adds published boolean field to twinkle_versions and makes full validation conditional on being published